fix: fix loader arguments
